latest commit

+++ blog_project/blog_app/views.py
@@ -1,6 +1,4 @@
-from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
-
 from .models import Blog
 
 
@@ -10,18 +8,21 @@ def home(request):
 
 
 def create_blog(request):
-    if request.method == 'POST':
-        title = request.POST['title']
-        body1 = request.POST['body1']
-        body2 = request.POST['body2']
-        body3 = request.POST['body3']
-        cover_image = request.FILES['coverimage']
-        subtitle1 = request.POST['subtitle1']
-        subtitle2 = request.POST['subtitle2']
-        blog = Blog.objects.create(title=title, body1=body1, body2=body2, body3=body3, author=request.user, cover_image=cover_image, subtitle1=subtitle1, subtitle2=subtitle2)
-        blog.save()
-        return redirect('my_blogs')
-    return render(request, "create_blog.html")
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            title = request.POST['title']
+            body1 = request.POST['body1']
+            body2 = request.POST['body2']
+            body3 = request.POST['body3']
+            cover_image = request.FILES['coverimage']
+            subtitle1 = request.POST['subtitle1']
+            subtitle2 = request.POST['subtitle2']
+            blog = Blog.objects.create(title=title, body1=body1, body2=body2, body3=body3, author=request.user, cover_image=cover_image, subtitle1=subtitle1, subtitle2=subtitle2)
+            blog.save()
+            return redirect('my_blogs')
+        return render(request, "create_blog.html")
+    else:
+        return redirect('signin')
 
 
 def view_blog(request, blog_id):
@@ -30,41 +31,52 @@ def view_blog(request, blog_id):
 
 
 def delete_blog(request, blog_id):
-    blog = Blog.objects.get(id=blog_id, author=request.user)
-    blog.delete()
-    return redirect('my_blogs')
-
+    if request.user.is_authenticated:
+        blog = Blog.objects.get(id=blog_id, author=request.user)
+        blog.delete()
+        return redirect('my_blogs')
+    else:
+        return redirect('signin')
 
 def my_blogs(request):
-    blog = Blog.objects.all().filter(author=request.user)
-    return render(request, 'my_blogs.html', {'blog': blog})
-
+    if request.user.is_authenticated:
+        blog = Blog.objects.all().filter(author=request.user)
+        return render(request, 'my_blogs.html', {'blog': blog})
+    else:
+        return redirect('signin')
 
 def view_my_blog(request, blog_id):
-    blog = Blog.objects.filter(id=blog_id)
-    return render(request, 'view_my_blog.html', {'blog': blog})
-
+    if request.user.is_authenticated:
+        blog = Blog.objects.filter(id=blog_id)
+        return render(request, 'view_my_blog.html', {'blog': blog})
+    else:
+        return redirect('signin')
 
 def update_blog(request, blog_id):
-    blog = Blog.objects.get(id=blog_id, author=request.user)
-    return render(request, 'update_blog.html', {'blog': blog})
-
+    if request.user.is_authenticated:
+        blog = Blog.objects.get(id=blog_id, author=request.user)
+        return render(request, 'update_blog.html', {'blog': blog})
+    else:
+        return redirect('signin')
 
 def update(request, blog_id):
-    title = request.POST['title']
-    coverimage = request.FILES['coverimage']
-    body = request.POST['body1']
-    subtitle1 = request.POST['subtitle1']
-    body2 = request.POST['body2']
-    subtitle2 = request.POST['subtitle2']
-    body3 = request.POST['body3']
-    blog = Blog.objects.get(id=blog_id, author=request.user)
-    blog.title = title
-    blog.cover_image = coverimage
-    blog.body1 = body
-    blog.subtitle1 = subtitle1
-    blog.body2 = body2
-    blog.subtitle2 = subtitle2
-    blog.body3 = body3
-    blog.save()
-    return redirect('my_blogs')
+    if request.user.is_authenticated:
+        title = request.POST['title']
+        coverimage = request.FILES['coverimage']
+        body = request.POST['body1']
+        subtitle1 = request.POST['subtitle1']
+        body2 = request.POST['body2']
+        subtitle2 = request.POST['subtitle2']
+        body3 = request.POST['body3']
+        blog = Blog.objects.get(id=blog_id, author=request.user)
+        blog.title = title
+        blog.cover_image = coverimage
+        blog.body1 = body
+        blog.subtitle1 = subtitle1
+        blog.body2 = body2
+        blog.subtitle2 = subtitle2
+        blog.body3 = body3
+        blog.save()
+        return redirect('my_blogs')
+    else:
+        return redirect('signin')
